+++ CHANGES.md
@@ -1,3 +1,4 @@
 v0.0.3 (dev)
 ============
 + fix bug for overlapping exons (#21)
++ prefer "22" over "22_KI270879v1_alt"